增加复合数据类型CVariedValue 增加消息组播功能，以位标记形式提供最多64个组，以及基于GroupID的组，前者组播效率较高 LuaWrap支持UNICODE，升级lua5.4，调整完善在c++层和lua层的对象管理 添加内置WebSocket支持 增加CEasyTimer64,CTimerQueue改为64位计时 调整了proxy的消息传递机制

+++ Libs/DBLib/DBTypes.h
@@ -224,7 +224,7 @@ typedef struct tagDB_COLUMN_INFO
 	char	Name[MAX_COLUMN_NAME];
 	short	Type;	
 	short	DigitSize;
-	UINT64	Size;
+	size_t	Size;
 }DB_COLUMN_INFO;
 
 
@@ -232,7 +232,7 @@ typedef struct tagDB_PARAM_INFO
 {	
 	short	Type;	
 	short	DigitSize;
-	UINT64	Size;
+	size_t	Size;
 	short	ParamType;
 	bool	IsNull;	
 }DB_PARAM_INFO;
